Move editable table types
