+++ src/components/editable-table/editable-table-collapsed-row.tsx
@@ -1,6 +1,6 @@
 import useStyles from './editable-table-collapsed-row.style'
 import { IconButton, TableCell, TableRow, Tooltip } from '@mui/material'
-import { TableDataRow } from '@/types/common'
+import { TableDataRow } from './types'
 import { EditableTableCell } from './editable-table-cell'
 import AddIcon from '@mui/icons-material/Add'
 import EditIcon from '@mui/icons-material/Edit'
