+++ src/components/editable-table/index.ts
@@ -1,0 +1,1 @@
+export * from './types'
