+++ src/components/editable-table/types.ts
@@ -1,0 +1,11 @@
+export type TableDataPoint = {
+  name: string
+  value: string
+  tooltip?: string
+  options?: string[] | undefined
+}
+
+export type TableDataRow = {
+  dataPoints: TableDataPoint[]
+  isNew: boolean
+}
